requirements.txt
argon2-cffi==23.1.0
authlib==1.5.1
black==24.8.0
crispy-bootstrap4==2024.1
Django==4.2.15
django-anymail==11.1
django-contact-form==5.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-maintenance-mode==0.21.1
django-oscar==3.2.5
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
pandas==2.2.2
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-pool==3.2.2
pycountry==24.6.1
pyjwt==2.10.1
pymysql==1.1.1
python-dotenv==1.0.1
sentry-sdk==2.13.0
sorl-thumbnail==12.10.0
weasyprint==62.3
google-api-python-client==2.141.0
google-genai
pyyaml
google-auth==2.35.0
google-auth-oauthlib==1.2.1
jsonschema==4.25.1